requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.5
django-allauth==0.33.0
django-cors-headers==2.1.0
django-extra-fields==0.9
django-rest-auth==0.9.1
djangorestframework==3.6.4
idna==2.6
oauthlib==2.0.4
olefile==0.44
Pillow==4.3.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
